Fix : Security 수정

+++ src/main/java/com/hallymfestival/HallymFestival2024BackEnd/domain/community/controller/CommunityAdminController.java
@@ -18,6 +18,7 @@ public class CommunityAdminController {
 
      private final CommunityService communityService;
 
+
      @GetMapping
      public ResponseEntity<List<CommunityEntity>> getCommunityList() {
         List<CommunityEntity> communityList = communityService.getCommunityList();
@@ -33,20 +34,29 @@ public class CommunityAdminController {
      }
 
      @DeleteMapping("/{id}")
-     @PreAuthorize("hasRole('admin')")
-     public ResponseEntity<CommunityResponseDto> removeCommunityAsAdmin(@PathVariable long id) {
-         CommunityResponseDto response = new CommunityResponseDto();
+     public ResponseEntity<CommunityResponseDto> removeCommunity(@PathVariable long id) {
+        // 1차 패스워드 일치 여부 확인.
+        boolean isOk = communityService.deleteAdminCommunity(id);
+        CommunityResponseDto response = new CommunityResponseDto();
 
-         boolean isDelete = communityService.deleteAdminCommunity(id);
+        if (!isOk) {
+            response.setCode("401");
+            response.setMessage("비밀번호가 일치하지 않습니다.");
+            return ResponseEntity.badRequest().body(response);
+        }
 
-         if (!isDelete) {
-             response.setCode("400");
-             response.setMessage("게시글 삭제에 실패하였습니다!");
-             return ResponseEntity.badRequest().body(response);
-         }
+        boolean isDelete = communityService.deleteAdminCommunity(id);
+        //boolean isDelete = true;
 
-         response.setCode("200");
-         response.setMessage("성공");
-         return ResponseEntity.ok().body(response);
+        if (!isDelete) {
+            response.setCode("400");
+            response.setMessage("게시글 삭제에 실패하였습니다.");
+            return ResponseEntity.badRequest().body(response);
+        }
+
+        response.setCode("200");
+        response.setMessage("성공");
+        return ResponseEntity.ok().body(response);
     }
+
 }
